[misc] Move humanize_size(), humanize_timestamp(), humanize_time_intervals() and get_shortcut_string() to util.py
